Do not use T1lib when compiling PHP
T1lib, used to support obsolete Type 1 fonts, was removed from Ubuntu 14.10. It seems that the t1lib site has been down and the downloads unavailable for a decade. I'm removing this dependency from our PHP build. The only "drawback" is that you won't be able to use Type 1 fonts. In reality, I've not used Type 1 fonts in any situation for more than 8 years so it's not a big loss after all.